feat(replaceglobalobjectvalues.ts): added the ability to convert stringified objects
refactored the fixNonStrings to work for stringified JSONs, and in so doing removed the need for the
code that converted stringified booleans and numbers.  A raw object still cannot be passed in, but a
stringified one will return the actual object.  In the process, wrapped fixNonStrings inside of
fixValueIfChanged, moved them to a separate file, and now fixValueIfChanged is called directly by
replaceGlobalObjectValues instead of replaceGlobalValuesInObject.